Add StepUnit class for recording changes
A StepUnit object holds information about a change being made to a cell.  Upon construction with this information, it also makes a string describing the change for the user.  Whenever a change is made to the puzzle, a StepUnit object should be added to a stack for record-keeping and the StepUnit.log_line should be added to Puzzle.solution_log.