Add Puzzle::check_for_none_available()
Use check_for_obvious_problems() for immediately after user initialization and use check_for_none_available() during the normal search.  (Since the program cannot allow an unavailable option to be written to a cell, row/column/house repeats are impossible problems in the normal search.  So the only thing we need to check is for unfilled cells with no available options.)